s01e04/requirements.txt
python-dotenv
requests
langchain-openai
langchain-core
